+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * Bug fix: only attempt to call `UIApplication.shared.open` on one URL at a time blocking other `open` calls until the current one is finished
   * Bug fix: Ensure that the `browserLoginAlertCanceled` event is not fired off when the `ASWebAutheniticationSession` popup is disabled.
   * Add `paypal:tokenize:browser-presentation:started` event for when the `ASWebAuthenticationSession.start` is launched.
+  * Update `url` FPTI tag to represent the approval URL prior to checkout and the success URL after checkout.
 
 ## 6.34.0 (2025-06-18)
 * BraintreePayPal